Rename Contacts namespace to Contracts namespace
