+++ stubs/Modules/Base/App/Contracts/IsSearchable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Contacts;
+namespace App\Contracts;
 
 interface IsSearchable
 {
