+++ stubs/Modules/Base/App/Models/Page.php
@@ -7,7 +7,7 @@ use App\Traits\HasVisibility;
 use App\Traits\Seoable;
 use App\Traits\Labelable;
 use App\Traits\Searchable;
-use App\Contacts\IsSearchable;
+use App\Contracts\IsSearchable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
