+++ stubs/Modules/Service/App/Models/Service.php
@@ -7,7 +7,7 @@ use Carbon\Carbon;
 use App\Traits\Seoable;
 use App\Traits\Labelable;
 use App\Traits\Searchable;
-use App\Contacts\IsSearchable;
+use App\Contracts\IsSearchable;
 use Awcodes\Curator\Models\Media;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
